requirements.txt
# College Voting System - Dependencies
# Web Framework
Flask==3.0.0
Flask-SQLAlchemy==3.1.1
Flask-Login==0.6.3
Flask-WTF==1.2.1
WTForms==3.1.1

# Database
SQLAlchemy==2.0.23

# Face Recognition
opencv-python==4.8.1.78
face_recognition==1.3.0
dlib==19.24.2
numpy==1.24.3
Pillow==10.1.0

# Utilities
python-dotenv==1.0.0
bcrypt==4.1.1
